// constant/type/error.type.ts
export type ErrorCoreService = {
  errorCode: string;
  errorMessageTh: string;
  errorMessage: string;
  errorDebugMessage: {
    message: string;
  };
};

export type ErrorTransFormAppResponse = {
  errorCode: string;
  errorMessageTh: string;
  errorMessage: string;
  errorDebugMessage: string;
};